FIX - Next.js eslint alerts
